pyproject.toml
[project]
name = "nonebot-plugin-pcrjjc"
version = "0.8.0"
authors = [
    {name="reine-ishyanami", email="2402979195@qq.com"}
]
description = "pcrjjc排名监测插件"
readme = "README.md"
license = { text = "AGPL-3" }
requires-python = ">=3.10, <4.0"
dependencies = [
    "msgpack>=1.0.5",
    "requests>=2.30.0",
    "Pillow>=9.5.0",
    "pydantic>=1.10.14",
    "python-dateutil>=2.8.2",
    "pycryptodome>=3.18.0",
    "nonebot2>=2.0.0",
    "nonebot-adapter-onebot>=2.2.3",
    "nonebot-plugin-apscheduler>=0.3.0"
]
classifiers = [
    "Framework :: Pydantic",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Natural Language :: Chinese (Simplified)"
]

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_apscheduler"]
plugin_dirs = ["src"]
builtin_plugins = []

[tool.setuptools.packages.find]
where = ["src"]
include = ["nonebot_plugin_pcrjjc"]

[build-system]
requires = ["setuptools >= 65.6.3"]
build-backend = "setuptools.build_meta"

[project.urls]
"Homepage" = "https://github.com/reine-ishyanami/nonebot-plugin-pcrjjc"
"Bug Tracker" = "https://github.com/reine-ishyanami/nonebot-plugin-pcrjjc/issues"